+++ tests/CMakeLists.txt
@@ -4,7 +4,7 @@ set(_test_targets
     )
 
 foreach (test ${_test_targets})
-  add_executable(${test} ${test}.cpp test_utility.cpp)
+  add_executable(${test} ${test}.cpp TestUtility.cpp)
   target_link_libraries(${test} PRIVATE JetsonGPIO)
 endforeach ()
 
@@ -16,7 +16,7 @@ set(_unit_test_targets
     )
 
 foreach (test ${_unit_test_targets})
-  add_executable(${test} unit-tests/${test}.cpp test_utility.cpp)
+  add_executable(${test} unit-tests/${test}.cpp TestUtility.cpp)
   add_test(NAME ${test} COMMAND $<TARGET_FILE:${test}>)
   target_link_libraries(${test} PRIVATE JetsonGPIO)
 endforeach ()
